[nvim] Fix toggling relativenumber on focus gain/loss
This fixes an annoying issue where re-entering an nvim window would
apparently jump the cursor into a floating pop-up window provided by the
LSP.

It turns out this is because of the 'windo' autocmd, which cycles
through all open windows and leaves focus on the last window. On nvim
this includes the floating windows, whereas on vim, the floating windows
aren't 'focusable' so they are skipped.

While I could probably configure coc/nvim to make the floating windows
non-focusable, there are some arguable benefits to them being focusable
in principle.

Instead, this commit ensures that the relativenumber behaviour doesn't
apply to floating windows, as it's not very useful on them anyway.